move workers system to separate package, fix handler registration
